fixed reviewer's remarks
