+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -1,14 +1,12 @@
 package ru.practicum.shareit.item;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.error.NotFoundException;
 import ru.practicum.shareit.error.ValidationException;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.item.service.ItemService;
+import ru.practicum.shareit.item.service.ItemServiceImpl;
 import ru.practicum.shareit.item.repository.ItemRepository;
-import ru.practicum.shareit.user.repository.UserRepository;
 
 import java.util.List;
 
@@ -19,17 +17,14 @@ import java.util.List;
 @RequestMapping("/items")
 public class ItemController {
 
-    final ItemService itemService;
-    final ItemRepository itemRepository;
-    final UserRepository userRepository;
-    final String EPItemId = "/{itemId}";
-    final String sharerId = "X-Sharer-User-Id";
+    final private ItemServiceImpl itemService;
+    final private ItemRepository itemRepository;
+    final private String epItemId = "/{itemId}";
+    final private String sharerId = "X-Sharer-User-Id";
 
-    @Autowired
-    public ItemController(ItemService itemService, ItemRepository itemRepository, UserRepository userRepository) {
+    public ItemController(ItemServiceImpl itemService, ItemRepository itemRepository) {
         this.itemService = itemService;
         this.itemRepository = itemRepository;
-        this.userRepository = userRepository;
     }
 
     @PostMapping
@@ -38,7 +33,7 @@ public class ItemController {
     }
 
 
-    @PatchMapping(EPItemId)
+    @PatchMapping(epItemId)
     public ItemDto updateItem(@RequestHeader(value = sharerId) Integer ownerId, @PathVariable Integer itemId, @RequestBody Item item) throws NotFoundException {
         return itemService.updateItem(itemId, item, ownerId, itemRepository);
     }
@@ -48,7 +43,7 @@ public class ItemController {
         return itemService.getAllItems(ownerId, itemRepository);
     }
 
-    @GetMapping(EPItemId)
+    @GetMapping(epItemId)
     public ItemDto getItemById(@PathVariable Integer itemId) {
         return itemService.getItemById(itemId, itemRepository);
     }
