+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -1,6 +1,5 @@
 package ru.practicum.shareit.user;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.error.ConflictException;
 import ru.practicum.shareit.error.NotFoundException;
@@ -8,7 +7,7 @@ import ru.practicum.shareit.error.ValidationException;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.repository.UserRepository;
-import ru.practicum.shareit.user.service.UserService;
+import ru.practicum.shareit.user.service.UserServiceImpl;
 
 import java.util.List;
 
@@ -19,12 +18,11 @@ import java.util.List;
 @RequestMapping(path = "/users")
 public class UserController {
 
-    final UserService userService;
-    final UserRepository userRepository;
-    final String EPUserId = "/{userId}";
+    final private UserServiceImpl userService;
+    final private UserRepository userRepository;
+    final private String epUserId = "/{userId}";
 
-    @Autowired
-    public UserController(UserService userService, UserRepository userRepository) {
+    public UserController(UserServiceImpl userService, UserRepository userRepository) {
         this.userService = userService;
         this.userRepository = userRepository;
     }
@@ -34,7 +32,7 @@ public class UserController {
         return userService.addUser(user, userRepository);
     }
 
-    @PatchMapping(EPUserId)
+    @PatchMapping(epUserId)
     public UserDto updateUser(@PathVariable int userId, @RequestBody User user) throws ValidationException, ConflictException {
         return userService.updateUser(userId, user, userRepository);
     }
@@ -44,12 +42,12 @@ public class UserController {
         return userService.getAllUsers(userRepository);
     }
 
-    @GetMapping(EPUserId)
+    @GetMapping(epUserId)
     public UserDto getUserById(@PathVariable int userId) throws NotFoundException {
         return userService.getUserById(userId, userRepository);
     }
 
-    @DeleteMapping(EPUserId)
+    @DeleteMapping(epUserId)
     public void deleteUser(@PathVariable int userId) {
         userService.deleteUser(userId, userRepository);
     }
